finish refactoring app.rs
