+++ src/app/mod.rs
@@ -7,18 +7,14 @@ use crate::cacher::Cacher;
 use crate::image_processor::ImageProcessor;
 use crate::lyrics_manager::LyricsManager;
 use crate::system_integration::SystemIntegration;
-use crate::worker_config::{WorkerConfig, calculate_worker_config};
 use crate::{
     audio::Audio,
     controller::{Controller, state::AppState},
     errors::AppError,
     scanner::Scanner,
-    ui::{
-        assets::Assets,
-        res_handler::{Event, ResHandler},
-        wiremann::Wiremann,
-    },
+    ui::{assets::Assets, res_handler::ResHandler, wiremann::Wiremann},
 };
+pub use paths::AppPaths;
 
 use gpui::{AppContext, Application, Result};
 use raw_window_handle::HasWindowHandle;
@@ -27,7 +23,7 @@ use std::sync::Arc;
 use events::{spawn_event_loop, subscribe_controller_events};
 use paths::{ensure_app_paths, get_app_paths};
 use window::build_window_options;
-use workers::spawn_worker;
+use workers::{WorkerConfig, calculate_worker_config, spawn_worker};
 
 static ICON_PNG: &[u8] = include_bytes!("../../assets/logos/logo.png");
 
@@ -50,8 +46,9 @@ pub fn run() -> Result<(), AppError> {
             ensure_app_paths(&app_paths);
 
             let app_icon = gpui::WindowIcon::from_png_bytes(ICON_PNG).ok();
+            let window_options = build_window_options(app_icon, cx);
 
-            cx.open_window(build_window_options(app_icon, cx), |window, cx| {
+            cx.open_window(window_options, |window, cx| {
                 let (mut audio, audio_tx, audio_rx) = Audio::new();
 
                 let (mut scanner, scanner_tx, scanner_rx) = Scanner::new(app_paths.clone());
@@ -108,7 +105,7 @@ pub fn run() -> Result<(), AppError> {
 
                 spawn_event_loop(cx, controller.clone(), arc_res.clone());
 
-                subscribe_controller_events(cx, &res_handler, controller.clone(), &view);
+                subscribe_controller_events(cx, &res_handler, controller.clone(), view.clone());
 
                 view
             })
