+++ src/controller/handlers/audio.rs
@@ -138,6 +138,4 @@ impl Controller {
         }
         Ok(())
     }
-
-
 }
